chore: better naming and error handling
