+++ crates/rome_cli/src/commands/check.rs
@@ -1,12 +1,25 @@
-use crate::commands::format::apply_format_settings_from_cli;
 use crate::configuration::load_configuration;
+use crate::parse_arguments::{apply_files_settings_from_cli, apply_format_settings_from_cli};
 use crate::{execute_mode, CliDiagnostic, CliSession, Execution, TraversalMode};
+use rome_console::{markup, ConsoleExt};
+use rome_diagnostics::PrintDiagnostic;
 use rome_service::workspace::{FixFileMode, UpdateSettingsParams};
 
 /// Handler for the "check" command of the Rome CLI
 pub(crate) fn check(mut session: CliSession) -> Result<(), CliDiagnostic> {
-    let mut configuration = load_configuration(&mut session)?;
-
+    let (mut configuration, diagnostics) = load_configuration(&mut session)?.consume();
+    if !diagnostics.is_empty() {
+        let console = &mut session.app.console;
+        console.log(markup!{
+           <Warn>"Found errors in the configuration file, Rome will use its defaults for the sections that are incorrect."</Warn>
+        });
+        for diagnostic in diagnostics {
+            console.log(markup! {
+                {PrintDiagnostic::verbose(&diagnostic)}
+            })
+        }
+    }
+    apply_files_settings_from_cli(&mut session, &mut configuration)?;
     apply_format_settings_from_cli(&mut session, &mut configuration)?;
 
     session
