+++ crates/rome_cli/src/commands/ci.rs
@@ -1,15 +1,28 @@
+use crate::parse_arguments::{apply_files_settings_from_cli, apply_format_settings_from_cli};
 use crate::{
     configuration::load_configuration, execute_mode, CliDiagnostic, CliSession, Execution,
     TraversalMode,
 };
+use rome_console::{markup, ConsoleExt};
+use rome_diagnostics::PrintDiagnostic;
 use rome_service::configuration::{FormatterConfiguration, LinterConfiguration};
 use rome_service::workspace::UpdateSettingsParams;
 
-use super::format::apply_format_settings_from_cli;
-
 /// Handler for the "ci" command of the Rome CLI
 pub(crate) fn ci(mut session: CliSession) -> Result<(), CliDiagnostic> {
-    let mut configuration = load_configuration(&mut session)?;
+    let (mut configuration, diagnostics) = load_configuration(&mut session)?.consume();
+
+    if !diagnostics.is_empty() {
+        let console = &mut session.app.console;
+        for diagnostic in diagnostics {
+            console.error(markup! {
+                {PrintDiagnostic::verbose(&diagnostic)}
+            })
+        }
+        return Err(CliDiagnostic::incompatible_end_configuration(
+            "The deserialization of the configuration resulted into an error.",
+        ));
+    }
 
     let formatter_enabled = session
         .args
@@ -42,6 +55,7 @@ pub(crate) fn ci(mut session: CliSession) -> Result<(), CliDiagnostic> {
         return Err(CliDiagnostic::incompatible_end_configuration("Formatter and Linter are both disabled, can't perform the command. This is probably and error."));
     }
 
+    apply_files_settings_from_cli(&mut session, &mut configuration)?;
     if !configuration.is_formatter_disabled() {
         apply_format_settings_from_cli(&mut session, &mut configuration)?;
     }
