+++ crates/rome_service/src/file_handlers/json.rs
@@ -1,11 +1,15 @@
 use super::{ExtensionHandler, Mime};
-use crate::file_handlers::{Capabilities, FormatterCapabilities, ParserCapabilities};
+use crate::file_handlers::{
+    AnalyzerCapabilities, Capabilities, FormatterCapabilities, LintParams, LintResults,
+    ParserCapabilities,
+};
 use crate::file_handlers::{DebugCapabilities, Language as LanguageId};
 use crate::settings::{
     FormatSettings, Language, LanguageSettings, LanguagesSettings, SettingsHandle,
 };
-use crate::workspace::GetSyntaxTreeResult;
-use crate::WorkspaceError;
+use crate::workspace::{GetSyntaxTreeResult, PullActionsResult};
+use crate::{Rules, WorkspaceError};
+use rome_diagnostics::{Diagnostic, Severity};
 use rome_formatter::{FormatError, Printed};
 use rome_fs::RomePath;
 use rome_json_formatter::context::JsonFormatOptions;
@@ -58,7 +62,12 @@ impl ExtensionHandler for JsonFileHandler {
                 debug_control_flow: None,
                 debug_formatter_ir: Some(debug_formatter_ir),
             },
-            analyzer: Default::default(),
+            analyzer: AnalyzerCapabilities {
+                lint: Some(lint),
+                code_actions: Some(code_actions),
+                rename: None,
+                fix_all: None,
+            },
             formatter: FormatterCapabilities {
                 format: Some(format),
                 format_range: Some(format_range),
@@ -162,4 +171,33 @@ fn format_on_type(
 
     let printed = rome_json_formatter::format_sub_tree(options, &root_node)?;
     Ok(printed)
+}
+
+fn lint(params: LintParams) -> LintResults {
+    let diagnostics = params.parse.into_diagnostics();
+
+    let diagnostic_count = diagnostics.len() as u64;
+    let errors = diagnostics
+        .iter()
+        .filter(|diag| diag.severity() <= Severity::Error)
+        .count();
+
+    let skipped_diagnostics = diagnostic_count - diagnostics.len() as u64;
+
+    LintResults {
+        diagnostics,
+        errors,
+        skipped_diagnostics,
+    }
+}
+
+fn code_actions(
+    _parse: AnyParse,
+    _range: TextRange,
+    _rules: Option<&Rules>,
+    _settings: SettingsHandle,
+) -> PullActionsResult {
+    PullActionsResult {
+        actions: Vec::new(),
+    }
 }
